Move Renderable into its own file.
